+++ packages/core/app-layout/sandbox/pages/SidebarPage.vue
@@ -49,16 +49,16 @@
         </div>
       </template>
       <template #sidebar-icon-overview>
-        <OverviewIcon :size="KUI_ICON_SIZE_40" />
+        <OverviewIcon :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`" />
       </template>
       <template #sidebar-icon-gateway-manager>
-        <RuntimesIcon :size="KUI_ICON_SIZE_40" />
+        <RuntimesIcon :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`" />
       </template>
       <template #sidebar-icon-servicehub>
-        <ServiceHubIcon :size="KUI_ICON_SIZE_40" />
+        <ServiceHubIcon :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`" />
       </template>
       <template #sidebar-icon-dev-portal>
-        <DevPortalIcon :size="KUI_ICON_SIZE_40" />
+        <DevPortalIcon :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`" />
       </template>
       <template #sidebar-after-dev-portal>
         <div class="new-badge-wrapper">
@@ -68,13 +68,13 @@
         </div>
       </template>
       <template #sidebar-icon-analytics>
-        <BarChartIcon :size="KUI_ICON_SIZE_40" />
+        <BarChartIcon :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`" />
       </template>
       <template #sidebar-icon-organization>
-        <PeopleIcon :size="KUI_ICON_SIZE_40" />
+        <PeopleIcon :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`" />
       </template>
       <template #sidebar-icon-settings>
-        <CogIcon :size="KUI_ICON_SIZE_40" />
+        <CogIcon :size="`var(--kui-icon-size-40, ${KUI_ICON_SIZE_40})`" />
       </template>
     </AppSidebar>
     <main>