+++ packages/entities/entities-plugins/src/components/free-form/shared/ArrayField.vue
@@ -175,7 +175,7 @@ import { useTemplateRef, nextTick, computed, ref, toValue, toRef, useAttrs } fro
 import { AddIcon, TrashIcon, CloseIcon } from '@kong/icons'
 import { KCard, type LabelAttributes } from '@kong/kongponents'
 import { replaceByDictionaryInFieldName, useField, useFieldAttrs, useFormShared, useItemKeys } from './composables'
-import useI18n from '../../../composables/useI18n'
+import useI18n from '../../../composables/useFreeformI18n'
 import * as utils from './utils'
 import Field from './Field.vue'
 import type { ArrayLikeFieldSchema } from '../../../types/plugins/form-schema'