+++ packages/entities/entities-plugins-icon/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins-icon@1.6.6...@kong-ui-public/entities-plugins-icon@1.6.7) (2026-06-11)
+
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies with stable versions ([#3435](https://github.com/Kong/public-ui-components/issues/3435)) ([5ef2562](https://github.com/Kong/public-ui-components/commit/5ef25621daf103979e3c399d4233cde0fac2d33d))
+
+
+
+
+
 ## [1.6.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins-icon@1.6.5...@kong-ui-public/entities-plugins-icon@1.6.6) (2026-05-27)